- Refactored get_licensed
