+++ tenable/io/__init__.py
@@ -20,7 +20,6 @@ Tenable.io
     assets
     audit_log
     credentials
-    display
     editor
     exclusions
     exports
@@ -59,7 +58,6 @@ from .assets import AssetsAPI
 from .audit_log import AuditLogAPI
 from .credentials import CredentialsAPI
 from .cs.api import ContainerSecurity
-from .display import DisplayAPI
 from .editor import EditorAPI
 from .exclusions import ExclusionsAPI
 from .exports.api import ExportsAPI
@@ -269,13 +267,6 @@ class TenableIO(APIPlatform):  # noqa: PLR0904
         return CredentialsAPI(self)
 
     @property
-    def display(self):
-        '''
-        The interface object for the Status and Account Info
-        '''
-        return DisplayAPI(self)
-
-    @property
     def editor(self):
         '''
         The interface object for the
@@ -450,4 +441,3 @@ class TenableIO(APIPlatform):  # noqa: PLR0904
         :doc:`Tenable.io v3 APIs <v3/index>`.
         '''
         return Version3API(self)
-
