+++ tests/io/test_workbenches.py
@@ -7,6 +7,7 @@ import pytest
 from tenable.errors import UnexpectedValueError, NotFoundError
 from tests.checker import check
 
+
 @pytest.mark.vcr()
 def test_workbench_assets_age_typeerror(api):
     '''
@@ -15,6 +16,7 @@ def test_workbench_assets_age_typeerror(api):
     with pytest.raises(TypeError):
         api.workbenches.assets(age='onetwothree')
 
+
 @pytest.mark.vcr()
 def test_workbench_assets_filter_tyype_typeerror(api):
     '''
@@ -22,6 +24,7 @@ def test_workbench_assets_filter_tyype_typeerror(api):
     '''
     with pytest.raises(TypeError):
         api.workbenches.assets(filter_type=1)
+
 
 @pytest.mark.vcr()
 def test_workbench_assets_filter_type_unexpectedvalueerror(api):
@@ -110,6 +113,7 @@ def test_workbench_assets(api):
         check(asset, 'terminated_by', str, allow_none=True)
         check(asset, 'updated_at', 'datetime')
 
+
 @pytest.mark.vcr()
 def test_workbench_assets_filtered(api):
     '''
@@ -190,6 +194,7 @@ def test_workbench_asset_activity(api):
     except NotFoundError:
         print('Activity for the asset uuid is not found')
 
+
 @pytest.mark.vcr()
 def test_workbench_asset_info_uuid_typeerror(api):
     '''
@@ -224,6 +229,7 @@ def test_workbench_asset_info(api):
     '''
     assets = api.workbenches.assets()
     api.workbenches.asset_info(assets[0]['id'])
+
 
 @pytest.mark.vcr()
 def test_workbench_asset_vulns_uuid_typeerror(api):
@@ -295,6 +301,7 @@ def test_workbench_asset_vulns(api):
             check(vulnerability, 'severity', int)
             check(vulnerability, 'vulnerability_state', str)
 
+
 @pytest.mark.vcr()
 def test_workbench_asset_vulns_filtered(api):
     '''
@@ -303,7 +310,7 @@ def test_workbench_asset_vulns_filtered(api):
     assets = api.workbenches.assets()
     if assets:
         vulns = api.workbenches.asset_vulns(assets[0]['id'],
-            ('severity', 'eq', 'Info'))
+                                            ('severity', 'eq', 'Info'))
         assert isinstance(vulns, list)
         if vulns:
             vulnerability = vulns[0]
@@ -318,6 +325,7 @@ def test_workbench_asset_vulns_filtered(api):
             check(vulnerability, 'plugin_name', str)
             check(vulnerability, 'severity', int)
             check(vulnerability, 'vulnerability_state', str)
+
 
 @pytest.mark.vcr()
 def test_workbench_asset_vuln_info_uuid_typeerror(api):
@@ -523,6 +531,7 @@ def test_workbench_asset_vuln_output(api):
                     check(result, 'severity', int)
                     check(result, 'transport_protocol', str)
 
+
 @pytest.mark.vcr()
 def test_workbench_vuln_assets(api):
     '''
@@ -546,6 +555,7 @@ def test_workbench_vuln_assets(api):
     for key, value in {'agent_name': list, 'last_seen': 'datetime', 'netbios_name': list}.items():
         if key in asset:
             check(asset, key, value)
+
 
 @pytest.mark.vcr()
 def test_workbench_export_asset_uuid_typeerror(api):
@@ -766,6 +776,7 @@ def test_workbench_vulns(api):
         check(vulnerability, 'recasted_count', int)
         check(vulnerability, 'vulnerability_state', str)
 
+
 @pytest.mark.vcr()
 def test_workbench_vuln_info_age_typeerror(api):
     '''
@@ -910,6 +921,7 @@ def test_workbench_vuln_outputs(api):
                     check(result, 'severity', int)
                     check(result, 'transport_protocol', str)
 
+
 @pytest.mark.vcr()
 def test_workbenches_asset_delete_asset_uuid_typeerror(api):
     '''
@@ -928,6 +940,117 @@ def test_workbenches_asset_delete_success(api):
     asset = api.workbenches.assets()[0]
     api.workbenches.asset_delete(asset['id'])
 
+
+@pytest.mark.vcr()
+def test_workbench_assets_fields(api):
+    '''
+    test to get the workbench assets and verifying their types
+    '''
+    assets = api.workbenches.assets(all_fields=True,
+                                    authenticated=True,
+                                    exploitable=True,
+                                    resolvable=True,
+                                    severity='critical')
+    assert isinstance(assets, list)
+    asset = assets[0]
+    check(asset, 'agent_name', list)
+    check(asset, 'aws_availability_zone', list)
+    check(asset, 'aws_ec2_instance_ami_id', list)
+    check(asset, 'aws_ec2_instance_group_name', list)
+    check(asset, 'aws_ec2_instance_id', list)
+    check(asset, 'aws_ec2_instance_state_name', list)
+    check(asset, 'aws_ec2_instance_type', list)
+    check(asset, 'aws_ec2_name', list)
+    check(asset, 'aws_ec2_product_code', list)
+    check(asset, 'aws_owner_id', list)
+    check(asset, 'aws_region', list)
+    check(asset, 'aws_subnet_id', list)
+    check(asset, 'aws_vpc_id', list)
+    check(asset, 'azure_resource_id', list)
+    check(asset, 'azure_vm_id', list)
+    check(asset, 'bios_uuid', list)
+    check(asset, 'created_at', 'datetime')
+    check(asset, 'first_scan_time', 'datetime', allow_none=True)
+    check(asset, 'first_seen', 'datetime', allow_none=True)
+    check(asset, 'fqdn', list)
+    check(asset, 'gcp_instance_id', list)
+    check(asset, 'gcp_project_id', list)
+    check(asset, 'gcp_zone', list)
+    check(asset, 'has_agent', bool)
+    check(asset, 'hostname', list)
+    check(asset, 'id', 'uuid')
+    check(asset, 'ipv4', list)
+    check(asset, 'ipv6', list)
+    check(asset, 'last_authenticated_scan_date', 'datetime', allow_none=True)
+    check(asset, 'last_licensed_scan_date', 'datetime', allow_none=True)
+    check(asset, 'last_scan_time', 'datetime', allow_none=True)
+    check(asset, 'last_seen', 'datetime', allow_none=True)
+    check(asset, 'mac_address', list)
+    check(asset, 'manufacturer_tpm_id', list)
+    check(asset, 'mcafee_epo_agent_guid', list)
+    check(asset, 'mcafee_epo_guid', list)
+    check(asset, 'netbios_name', list)
+    check(asset, 'operating_system', list)
+    check(asset, 'qualys_asset_id', list)
+    check(asset, 'qualys_host_id', list)
+    check(asset, 'servicenow_sysid', list)
+    check(asset, 'sources', list)
+    for source in asset['sources']:
+        check(source, 'first_seen', 'datetime')
+        check(source, 'last_seen', 'datetime')
+        check(source, 'name', str)
+    check(asset, 'ssh_fingerprint', list)
+    check(asset, 'symantec_ep_hardware_key', list)
+    check(asset, 'system_type', list)
+    check(asset, 'tags', list)
+    check(asset, 'tenable_uuid', list)
+    check(asset, 'terminated_at', 'datetime', allow_none=True)
+    check(asset, 'terminated_by', str, allow_none=True)
+    check(asset, 'updated_at', 'datetime')
+
+
+@pytest.mark.vcr()
+def test_workbenches_export_success(api):
+    '''
+    test to export the data from vulnerability workbench
+    '''
+    with open('example.nessus', 'wb') as fobj:
+        api.workbenches.export(filter_type='or',
+                               fobj=fobj)
+
+
+@pytest.mark.vcr()
+def test_workbench_vulns_fields(api):
+    '''
+    test to get the workbench vulnerabilities
+    '''
+    vulns = api.workbenches.vulns(authenticated=True,
+                                  severity='medium',
+                                  exploitable=True,
+                                  resolvable=True)
+    assert isinstance(vulns, list)
+    each_vuln = vulns[0]
+    check(each_vuln, 'accepted_count', int)
+    check(each_vuln, 'counts_by_severity', list)
+    for data in each_vuln['counts_by_severity']:
+        check(data, 'count', int)
+        check(data, 'value', int)
+    check(each_vuln, 'plugin_family', str)
+    check(each_vuln, 'plugin_id', int)
+    check(each_vuln, 'plugin_name', str)
+    check(each_vuln, 'recasted_count', int)
+    check(each_vuln, 'vulnerability_state', str)
+
+
+@pytest.mark.vcr()
+def test_workbench_asset_info_success(api):
+    '''
+    test to get the asset information
+    '''
+    asset = api.workbenches.assets()
+    api.workbenches.asset_info(asset[0]['id'], all_fields=False)
+
+
 @pytest.mark.vcr()
 def test_workbench_assets_fields(api):
     '''
@@ -1038,112 +1161,8 @@ def test_workbench_asset_info_success(api):
     api.workbenches.asset_info(asset[0]['id'], all_fields=False)
 
 
-
 @pytest.mark.vcr()
-def test_workbench_assets_fields(api):
-    '''
-    test to get the workbench assets and verifying their types
-    '''
-    assets = api.workbenches.assets(all_fields=True,
-                                    authenticated=True,
-                                    exploitable=True,
-                                    resolvable=True,
-                                    severity='critical')
-    assert isinstance(assets, list)
-    asset = assets[0]
-    check(asset, 'agent_name', list)
-    check(asset, 'aws_availability_zone', list)
-    check(asset, 'aws_ec2_instance_ami_id', list)
-    check(asset, 'aws_ec2_instance_group_name', list)
-    check(asset, 'aws_ec2_instance_id', list)
-    check(asset, 'aws_ec2_instance_state_name', list)
-    check(asset, 'aws_ec2_instance_type', list)
-    check(asset, 'aws_ec2_name', list)
-    check(asset, 'aws_ec2_product_code', list)
-    check(asset, 'aws_owner_id', list)
-    check(asset, 'aws_region', list)
-    check(asset, 'aws_subnet_id', list)
-    check(asset, 'aws_vpc_id', list)
-    check(asset, 'azure_resource_id', list)
-    check(asset, 'azure_vm_id', list)
-    check(asset, 'bios_uuid', list)
-    check(asset, 'created_at', 'datetime')
-    check(asset, 'first_scan_time', 'datetime', allow_none=True)
-    check(asset, 'first_seen', 'datetime', allow_none=True)
-    check(asset, 'fqdn', list)
-    check(asset, 'gcp_instance_id', list)
-    check(asset, 'gcp_project_id', list)
-    check(asset, 'gcp_zone', list)
-    check(asset, 'has_agent', bool)
-    check(asset, 'hostname', list)
-    check(asset, 'id', 'uuid')
-    check(asset, 'ipv4', list)
-    check(asset, 'ipv6', list)
-    check(asset, 'last_authenticated_scan_date', 'datetime', allow_none=True)
-    check(asset, 'last_licensed_scan_date', 'datetime', allow_none=True)
-    check(asset, 'last_scan_time', 'datetime', allow_none=True)
-    check(asset, 'last_seen', 'datetime', allow_none=True)
-    check(asset, 'mac_address', list)
-    check(asset, 'manufacturer_tpm_id', list)
-    check(asset, 'mcafee_epo_agent_guid', list)
-    check(asset, 'mcafee_epo_guid', list)
-    check(asset, 'netbios_name', list)
-    check(asset, 'operating_system', list)
-    check(asset, 'qualys_asset_id', list)
-    check(asset, 'qualys_host_id', list)
-    check(asset, 'servicenow_sysid', list)
-    check(asset, 'sources', list)
-    for source in asset['sources']:
-        check(source, 'first_seen', 'datetime')
-        check(source, 'last_seen', 'datetime')
-        check(source, 'name', str)
-    check(asset, 'ssh_fingerprint', list)
-    check(asset, 'symantec_ep_hardware_key', list)
-    check(asset, 'system_type', list)
-    check(asset, 'tags', list)
-    check(asset, 'tenable_uuid', list)
-    check(asset, 'terminated_at', 'datetime', allow_none=True)
-    check(asset, 'terminated_by', str, allow_none=True)
-    check(asset, 'updated_at', 'datetime')
-
-
-@pytest.mark.vcr()
-def test_workbenches_export_success(api):
-    '''
-    test to export the data from vulnerability workbench
-    '''
-    with open('example.nessus', 'wb') as fobj:
-        api.workbenches.export(filter_type='or',
-                               fobj=fobj)
-
-
-@pytest.mark.vcr()
-def test_workbench_vulns_fields(api):
-    '''
-    test to get the workbench vulnerabilities
-    '''
-    vulns = api.workbenches.vulns(authenticated=True,
-                                  severity='medium',
-                                  exploitable=True,
-                                  resolvable=True)
-    assert isinstance(vulns, list)
-    each_vuln = vulns[0]
-    check(each_vuln, 'accepted_count', int)
-    check(each_vuln, 'counts_by_severity', list)
-    for data in each_vuln['counts_by_severity']:
-        check(data, 'count', int)
-        check(data, 'value', int)
-    check(each_vuln, 'plugin_family', str)
-    check(each_vuln, 'plugin_id', int)
-    check(each_vuln, 'plugin_name', str)
-    check(each_vuln, 'recasted_count', int)
-    check(each_vuln, 'vulnerability_state', str)
-
-
-@pytest.mark.vcr()
-def test_workbench_asset_info_success(api):
-    '''
-    test to get the asset information
-    '''
-    asset = api.workbenches.assets()
-    api.workbenches.asset_info(asset[0]['id'], all_fields=False)
+def test_get_licensed(api):
+    """test to get licensed scanned assets count"""
+    data = api.workbenches.get_licensed()
+    assert data == 2502
